Refactor transversal themes parsing, multiple fixes
Rewritten parse to account for the new structure of the TT sheet in the
input data.

Several fixes in the input spreadhseet. Main ones:

* All themes have a second level, even if they only have one category
* Replace codes with the actual indicator code (eg 3.3.1a -> 3.3.1_gp)
* Capitalize titles for better display

Also added placeholders for missing descriptions.

TODO: What to do with calculated indicators eg parity indicators that
result from dividing 2 of the others. Proposed solution is to add them
as proper indicators on the main sheeet.